fetchDenoDeps: rename fetch-default to fetch-common
